requirements.txt
chainlit==1.0.200
openai==0.28.0  # 最新バージョンに更新
python-dotenv==1.0.0  # .envファイルからAPIキーを読み込むため
faiss-cpu
numpy
pdfplumber
PyPDF2
